Move tests to BulkActions folder and namespace
